requirements-test.txt
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
pytest-timeout>=2.1.0

torch>=2.0.0
transformers>=4.45.2
numpy>=1.21.0
scikit-learn>=1.0.0
pandas>=1.3.0
pyyaml>=6.0


bitsandbytes>=0.41.0
accelerate>=0.20.0


hdbscan>=0.8.29
matplotlib>=3.5.0
seaborn>=0.11.0


coverage>=6.0
pytest-cov>=4.0
black>=22.0
isort>=5.0
flake8>=4.0

# vllm